Fix code locations
