--- src/Microsoft.TemplateEngine.Authoring.TemplateVerifier/Expectations/made-up-template.--a#-b#c#--d.Windows.verified/std-streams/stdout.txt
@@ -1,1 +1,0 @@
-﻿stdout content
