--- src/Microsoft.TemplateEngine.Authoring.TemplateVerifier/PhysicalFileSystem.cs
@@ -1,21 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace Microsoft.TemplateEngine.Authoring.TemplateVerifier
-{
-    internal class PhysicalFileSystem : IFileSystem
-    {
-        public bool DirectoryExists(string path) => Directory.Exists(path);
-
-        public IEnumerable<string> EnumerateFiles(string path, string searchPattern, SearchOption searchOption) =>
-            Directory.EnumerateFiles(path, searchPattern, searchOption);
-
-        public Task<string> ReadAllTextAsync(string path) => File.ReadAllTextAsync(path);
-
-        public Task WriteAllTextAsync(string path, string? contents) => File.WriteAllTextAsync(path, contents);
-
-        public void CreateDirectory(string path) => Directory.CreateDirectory(path);
-
-        public void DeleteDirectory(string path, bool recursive) => Directory.Delete(path, recursive);
-    }
-}
