+++ test/Microsoft.TemplateEngine.Authoring.TemplateVerifier.IntegrationTests/ExampleTemplateTest.cs
@@ -1,0 +1,48 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Microsoft.Extensions.Logging;
+using Microsoft.TemplateEngine.TestHelper;
+using Microsoft.TemplateEngine.Tests;
+using Xunit.Abstractions;
+
+namespace Microsoft.TemplateEngine.Authoring.TemplateVerifier.IntegrationTests
+{
+    public class ExampleTemplateTest : TestBase
+    {
+        private readonly ILogger _log;
+
+        public ExampleTemplateTest(ITestOutputHelper log)
+        {
+            _log = new XunitLoggerProvider(log).CreateLogger("TestRun");
+        }
+
+        [Fact]
+        public async void VerificationEngineSampleDogfoodTest()
+        {
+            string workingDir = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName().Replace(".", string.Empty));
+            string templateShortName = "TestAssets.SampleTestTemplate";
+
+            //get the template location
+            string executingAssemblyPath = this.GetType().Assembly.Location;
+            string templateLocation = Path.Combine(TestTemplatesLocation, "TestTemplate");
+
+            TemplateVerifierOptions options = new TemplateVerifierOptions(templateName: templateShortName)
+            {
+                TemplateSpecificArgs = new string[] { "--paramB", "true" },
+                TemplatePath = templateLocation,
+                ExpectationsDirectory = "Expectations",
+                OutputDirectory = workingDir,
+                VerifyCommandOutput = true,
+                UniqueFor = UniqueForOption.RuntimeAndVersion,
+            }
+                .WithCustomScrubbers(
+                    ScrubbersDefinition.Empty
+                        .AddScrubber(sb => sb.Replace("B is enabled", "*******"))
+                );
+
+            VerificationEngine engine = new VerificationEngine(_log);
+            await engine.Execute(options);
+        }
+    }
+}
