+++ test/Microsoft.TemplateEngine.Authoring.TemplateVerifier.IntegrationTests/Microsoft.TemplateEngine.Authoring.TemplateVerifier.IntegrationTests.csproj
@@ -8,16 +8,17 @@
   <ItemGroup>
     <ProjectReference Include="$(SrcDir)Microsoft.TemplateEngine.Authoring.TemplateVerifier\Microsoft.TemplateEngine.Authoring.TemplateVerifier.csproj" />
     <ProjectReference Include="$(TestDir)Microsoft.TemplateEngine.TestHelper\Microsoft.TemplateEngine.TestHelper.csproj" />
+    <ProjectReference Include="..\Microsoft.TemplateEngine.TestTemplates\Microsoft.TemplateEngine.TestTemplates.csproj" />
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Remove="TestTemplate\*" />
-    <Content Include="TestTemplate\**\*" Exclude="TestTemplate\**\bin\**;TestTemplate\**\obj\**" />
-    <None Include="TestTemplate\**\*" CopyToOutputDirectory="PreserveNewest" />
-
     <Compile Remove="Expectations\**\*" />
     <Content Include="Expectations\**\*" />
     <None Include="Expectations\**\*" CopyToOutputDirectory="Never" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\Shared\**\*.cs" Link="%(RecursiveDir)%(Filename)%(Extension)" />
   </ItemGroup>
 
 </Project>
