--- test/Microsoft.TemplateEngine.Authoring.TemplateVerifier.IntegrationTests/TestTemplate/.template.config/template.json
@@ -1,24 +1,0 @@
-{
-  "author": "Test Asset",
-  "classifications": [ "Test Asset" ],
-  "name": "SampleTestTemplate",
-  "generatorVersions": "[1.0.0.0-*)",
-  "groupIdentity": "TestAssets.SampleTestTemplate",
-  "precedence": "100",
-  "identity": "TestAssets.SampleTestTemplate",
-  "shortName": "TestAssets.SampleTestTemplate",
-  "symbols": {
-    "paramA": {
-      "type": "parameter",
-      "datatype": "string",
-      "replaces": "placeholderA",
-      "defaultValue":  "false" 
-    },
-    "paramB": {
-      "type": "parameter",
-      "datatype": "string",
-      "isRequired": true,
-      "replaces": "placeholderB"
-    }
-  }
-}
