--- test/Microsoft.TemplateEngine.Authoring.TemplateVerifier.IntegrationTests/TestTemplate/Test.cs
@@ -1,11 +1,0 @@
-
-// value of paramA: placeholderA
-// value of paramB: placeholderB
-
-//#if( paramA )
-	// A is enabled
-//#endif
-
-//#if( paramB )
-	// B is enabled
-//#endif
