+++ test/Microsoft.TemplateEngine.TestTemplates/test_templates/TestTemplate/.template.config/template.json
